style(subway): rename the initialization class
- 지하철 프로그램을 초기 설정하는 클래스 이름 변경
